fix(NavItem): fix NavItem cannot be used in Navbar and Sidenav
